pyproject.toml
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pytest.ini_options]
testpaths = "tests/"
timeout = 45
filterwarnings = [
  "ignore::DeprecationWarning"
]
asyncio_default_fixture_loop_scope = "session"

[tool.pylint.main]
disable = [
  "broad-exception-caught",
  "broad-exception-raised",
  "disallowed-name",
  "fixme",
  "invalid-name",
  "logging-fstring-interpolation",
  "missing-class-docstring",
  "missing-function-docstring",
  "pointless-string-statement",
  "possibly-used-before-assignment",
  "redefined-builtin",
  "too-few-public-methods",
  "too-many-arguments",
  "too-many-branches",
  "too-many-instance-attributes",
  "too-many-lines",
  "too-many-locals",
  "too-many-positional-arguments",
  "too-many-public-methods",
  "too-many-statements",
  "unspecified-encoding",
]
jobs = 4

[tool.black]
line-length = 100
target-version = ['py311']

[project]
authors = [
    {name = "Paul Kang"},
    {name = "Kyeyoon Park"},
]
license = {text = "BSD-3"}
requires-python = "<4.0,>=3.11"
dependencies = [
    "click>=8.1.7",
    "humanfriendly>=10.0",
    "pyyaml>=6.0.1",
    "python-socketio>=5.10.0",
    "aiohttp>=3.8.6",
    "pytest-timeout>=2.2.0",
    "dill>=0.3.7",
    "websockets>=12.0",
    "jsonrpcserver>=5.0.9",
    "jsonrpcclient>=4.0.3",
    "black>=24.4.2",
    "sortedcontainers>=2.4.0",
    "readerwriterlock>=1.0.9",
    "torch>=2.3.1",
    "torchvision>=0.18.1",
    "torchinfo>=1.8.0",
    "tqdm>=4.66.4",
    "pyshark>=0.6",
    "scapy>=2.6.1",
    "pytest-xdist>=3.5.0",
    "fastapi>=0.115.12",
    "uvicorn>=0.34.2",
    "langchain>=0.3.25",
    "langchain-community>=0.3.24",
    "langchain-ollama>=0.3.3",
    "python-multipart>=0.0.20",
    "pypdf>=5.5.0",
    "sentence-transformers>=4.1.0",
    "faiss-cpu>=1.11.0",
    "langchain-huggingface>=0.2.0",
]
name = "opencis"
version = "0.4.0"
description = "OpenCXL"
readme = "README.md"

[dependency-groups]
dev = [
    "pytest>=7.3.1",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.1",
    "pylint>=3.0.1",
]
